Add lua script to support atomic operation in redis server

+++ DistributedRateLimiter/Configuration/RateLimiterOptions.cs
@@ -4,7 +4,7 @@ public class RateLimiterOptions
 {
     public string Name { get; set; } = nameof(RateLimiter);
 
-    public double BurstCapacity { get; set; }
+    public int BurstCapacity { get; set; }
 
     public int RefillPerSecond { get; set; }
 }
